refactor: move all function defs to core
