+++ packages/core/package.json
@@ -58,6 +58,7 @@
     "definition": "dist/index.d.ts"
   },
   "dependencies": {
+    "@formatjs/icu-messageformat-parser": "^2.11.4",
     "crypto-js": "^4.2.0",
     "fast-json-stable-stringify": "^2.1.0",
     "intl-messageformat": "^10.7.16"
