+++ packages/core/src/static/index.ts
@@ -1,0 +1,4 @@
+export * from './decodeVars';
+export * from './declareVar';
+export * from './declareStatic';
+export * from './indexVars';
