+++ packages/core/src/static/indexVars.ts
