+++ packages/core/src/static/utils/__tests__/sanitizeVar.test.ts
