+++ packages/core/src/static/utils/constants.ts
