+++ packages/core/src/static/utils/sanitizeVar.ts
