+++ packages/core/src/static/utils/traverseIcu.ts
