+++ packages/i18n/src/index.ts
@@ -1,1 +1,2 @@
 export * from './msg';
+export * from './static';
