+++ packages/i18n/src/static/index.ts
@@ -1,0 +1,5 @@
+export {
+  declareStatic,
+  declareVar,
+  decodeVars,
+} from 'generaltranslation/internal';
