+++ packages/next/src/index.client.ts
@@ -6,7 +6,6 @@ import {
   Currency,
   DateTime,
   Static,
-  declareStatic,
   T,
   Branch,
   Plural,
@@ -24,6 +23,9 @@ import {
   msg,
   decodeMsg,
   decodeOptions,
+  declareStatic,
+  declareVar,
+  decodeVars,
   mFallback,
   gtFallback,
 } from 'gt-react/client';
@@ -54,7 +56,6 @@ export {
   Currency,
   DateTime,
   Static,
-  declareStatic,
   Branch,
   Plural,
   LocaleSelector,
@@ -74,6 +75,9 @@ export {
   msg,
   decodeMsg,
   decodeOptions,
+  declareStatic,
+  declareVar,
+  decodeVars,
   mFallback,
   gtFallback,
 };
