+++ packages/react-core/src/index.ts
@@ -9,7 +9,7 @@ import Var from './variables/Var';
 import Num from './variables/Num';
 import Currency from './variables/Currency';
 import DateTime from './variables/DateTime';
-import { Static, declareStatic } from './variables/Static';
+import { Static } from './variables/Static';
 import Plural from './branches/plurals/Plural';
 import Branch from './branches/Branch';
 import useLocales from './hooks/useLocales';
@@ -35,13 +35,14 @@ import { useCreateInternalUseTranslationsObjFunction } from './provider/hooks/tr
 
 export * from 'gt-i18n/fallbacks';
 
+export { declareStatic, declareVar, decodeVars } from 'gt-i18n';
+
 export {
   Var,
   Num,
   Currency,
   DateTime,
   Static,
-  declareStatic,
   T,
   GTProvider,
   Plural,
