+++ packages/react-core/src/internal.ts
@@ -33,9 +33,11 @@ import { injectFallbacks } from './dictionaries/injectFallbacks';
 import { injectAndMerge } from './dictionaries/injectAndMerge';
 import { collectUntranslatedEntries } from './dictionaries/collectUntranslatedEntries';
 import { msg, decodeMsg, decodeOptions } from './messages/messages';
-import { Static, declareStatic } from './variables/Static';
+import { Static } from './variables/Static';
 
 export * from 'gt-i18n/fallbacks';
+export { declareStatic, declareVar, decodeVars } from 'gt-i18n';
+
 export {
   addGTIdentifier,
   writeChildrenAsObjects,
@@ -70,5 +72,4 @@ export {
   decodeMsg,
   decodeOptions,
   Static,
-  declareStatic,
 };
