+++ packages/react/src/client.ts
@@ -15,7 +15,6 @@ import {
   Currency,
   DateTime,
   Static,
-  declareStatic,
   Plural,
   Branch,
   useLocales,
@@ -29,6 +28,9 @@ import {
   msg,
   decodeMsg,
   decodeOptions,
+  declareStatic,
+  declareVar,
+  decodeVars,
   mFallback,
   gtFallback,
 } from '@generaltranslation/react-core';
@@ -51,6 +53,9 @@ export {
   msg,
   decodeMsg,
   decodeOptions,
+  declareStatic,
+  declareVar,
+  decodeVars,
   useMessages,
   mFallback,
   gtFallback,
@@ -60,7 +65,6 @@ export {
   DateTime,
   Currency,
   Static,
-  declareStatic,
   Branch,
   Plural,
   LocaleSelector,
