+++ packages/react/src/index.ts
@@ -14,7 +14,6 @@ import {
   Currency,
   DateTime,
   Static,
-  declareStatic,
   Plural,
   Branch,
   useLocales,
@@ -28,6 +27,9 @@ import {
   msg,
   decodeMsg,
   decodeOptions,
+  declareStatic,
+  declareVar,
+  decodeVars,
   mFallback,
   gtFallback,
 } from '@generaltranslation/react-core';
@@ -43,7 +45,6 @@ export {
   Currency,
   DateTime,
   Static,
-  declareStatic,
   T,
   GTProvider,
   Plural,
@@ -68,6 +69,9 @@ export {
   msg,
   decodeMsg,
   decodeOptions,
+  declareStatic,
+  declareVar,
+  decodeVars,
   useMessages,
   mFallback,
   gtFallback,
